requirements.txt
opencv-python-headless
pytesseract
groq
python-dotenv
flask
werkzeug
gunicorn
httpx
